Moved PasswordInput under Inputs folder
