+++ app/components/Inputs/PasswordInput/index.js
@@ -1,0 +1,1 @@
+export { default } from './PasswordInput'
